insulin: use event-based architecture components

+++ insulin/src/main/java/it/diab/insulin/fragments/EditorFragment.kt
@@ -12,36 +12,28 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ArrayAdapter
-import android.widget.TextView
-import androidx.appcompat.widget.AppCompatEditText
-import androidx.appcompat.widget.AppCompatSpinner
-import androidx.appcompat.widget.SwitchCompat
 import androidx.lifecycle.ViewModelProviders
-import com.google.android.material.button.MaterialButton
 import it.diab.core.util.Activities
+import it.diab.core.util.extensions.bus
 import it.diab.data.entities.Insulin
 import it.diab.data.entities.TimeFrame
 import it.diab.data.repositories.InsulinRepository
 import it.diab.insulin.R
+import it.diab.insulin.components.EditableComponent
+import it.diab.insulin.components.EditorActionsComponent
+import it.diab.insulin.components.status.EditableInStatus
+import it.diab.insulin.events.EditEvent
 import it.diab.insulin.viewmodels.EditorViewModel
 import it.diab.insulin.viewmodels.EditorViewModelFactory
-import it.diab.ui.util.UIUtils
 import it.diab.ui.widgets.BottomSheetDialogFragmentExt
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 
 class EditorFragment : BottomSheetDialogFragmentExt() {
 
-    private lateinit var titleView: TextView
-    private lateinit var editText: AppCompatEditText
-    private lateinit var spinner: AppCompatSpinner
-    private lateinit var basalSwitch: SwitchCompat
-    private lateinit var halfUnitsSwitch: SwitchCompat
-    private lateinit var saveButton: MaterialButton
-    private lateinit var deleteButton: MaterialButton
-
     private lateinit var viewModel: EditorViewModel
 
-    private var editMode = false
+    private val uiScope = CoroutineScope(Dispatchers.Main)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -55,77 +47,42 @@ class EditorFragment : BottomSheetDialogFragmentExt() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        val view = inflater.inflate(R.layout.fragment_insulin_edit, container, false)
-
-        titleView = view.findViewById(R.id.insulin_edit_dialog_title)
-        editText = view.findViewById(R.id.insulin_edit_name)
-        spinner = view.findViewById(R.id.insulin_edit_time)
-        basalSwitch = view.findViewById(R.id.insulin_edit_basal)
-        halfUnitsSwitch = view.findViewById(R.id.insulin_edit_half_units)
-        saveButton = view.findViewById(R.id.insulin_edit_save)
-        deleteButton = view.findViewById(R.id.insulin_edit_delete)
-
-        UIUtils.setWhiteNavBarIfNeeded(requireContext(), dialog)
-
-        return view
-    }
+    ): View = inflater.inflate(R.layout.fragment_insulin_edit, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        EditableComponent(view, uiScope, bus)
+        EditorActionsComponent(view, bus)
 
         val uid = arguments?.getLong(Activities.Insulin.EXTRA_EDITOR_UID, -1L) ?: -1L
-        editMode = uid >= 0
         viewModel.setInsulin(uid, this::setup)
     }
 
     private fun setup(insulin: Insulin) {
-
-        titleView.text = getString(if (editMode) R.string.insulin_editor_edit else R.string.insulin_editor_add)
-
-        editText.setText(insulin.name)
-        basalSwitch.isChecked = insulin.isBasal
-        halfUnitsSwitch.isChecked = insulin.hasHalfUnits
-
-        setupSpinner(insulin)
-
-        saveButton.setOnClickListener { saveInsulin() }
-        deleteButton.setOnClickListener { deleteInsulin() }
-
-        if (!editMode) {
-            deleteButton.visibility = View.GONE
+        bus.subscribe(EditEvent::class, uiScope) {
+            when (it) {
+                is EditEvent.IntentSave -> {
+                    viewModel.save(it.status)
+                    dismiss()
+                }
+                is EditEvent.IntentRequestDelete -> {
+                    viewModel.delete()
+                    dismiss()
+                }
+            }
         }
-    }
 
-    private fun setupSpinner(insulin: Insulin) {
-        val context = context ?: return
-
-        val timeFrames = TimeFrame.values()
-        spinner.apply {
-            adapter = ArrayAdapter(
-                context,
-                androidx.appcompat.R.layout.support_simple_spinner_dropdown_item,
-                timeFrames.map { timeFrame -> getString(timeFrame.string) }
+        bus.emit(
+            EditEvent::class, EditEvent.IntentEdit(
+                EditableInStatus(
+                    insulin.uid > 0L,
+                    insulin.name,
+                    insulin.timeFrame.toInt() + 1,
+                    TimeFrame.values().map(TimeFrame::string),
+                    insulin.hasHalfUnits,
+                    insulin.isBasal
+                )
             )
-
-            setSelection(insulin.timeFrame.toInt() + 1)
-        }
-    }
-
-    private fun saveInsulin() {
-        viewModel.insulin.apply {
-            name = editText.text.toString()
-            isBasal = basalSwitch.isChecked
-            hasHalfUnits = halfUnitsSwitch.isChecked
-            setTimeFrame(spinner.selectedItemPosition - 1)
-        }
-
-        viewModel.save()
-        dismiss()
-    }
-
-    private fun deleteInsulin() {
-        viewModel.delete()
-        dismiss()
+        )
     }
 }
